Restructure scene management.
There's a bit of conflict between the way scenes are designed
and how Apple intends for you to switch between scenes.
In order to keep a scene in memory, you will need to keep
it in the UIViewController class, and scenes have to keep
a pointer to the UIVC, I don't particularly like it but
it seems as though there isn't much I can do otherwise.