Add Pacer pattern, update animations
Small distinction between Barrage and Pacer patterns, a Barrage is a
non-stop sequence of barriers, a Pacer is considered 'normal' gameplay
where it's a combination of everything.

I also added fade-out animations to barriers and modifiers so it's
clearer when you interact with them.  The player is now a small gold
circle (still temporary).