Undo previous commit changes, and cleanup unused sounds
Now I remember why it was like that, the SKAudioNode was used for the
looping music track, and to start and end that you need to addChild and
removeFromParent (gross).  We will replace this, so I just put a little
comment to make it a little more apparent what it's doing.