Change sequencer to generate entire level upfront
This is more convenient than generating entities on the fly.  However, I
think we're going to run into sync errors if we try and start the level
anywhere other than at the beginning.  Maybe if we pass in a Random
object with a known seed it will generate the same patterns over and
over again.